requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
pandas==2.1.3
python-pptx==0.6.23
jinja2==3.1.2
Markdown==3.5.1
pydantic==2.5.0
aiohttp==3.9.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1 